// src/libs/types/pokemon.type.ts
export interface Result {
  name: string;
  url?: string;
}

export interface ListItem {
  key: string;
  value: string | never;
  url?: string;
}

export type PokemonListRes = {
  count?: number;
  next?: string;
  previous?: string;
  results?: Result[];
  pokemon?: Result[];
};

interface ErrorHandling {
  success?: boolean;
  error?: { message: string };
}

export type ResultsMOckApi = { results: Result[] } & ErrorHandling;

export type PokemonListApiRes = PokemonListRes & ErrorHandling;

// Search Pokemon: https://pokeapi.co/api/v2/pokemon/{id-or-name}/

export interface Ability2 {
  name: string;
  url: string;
}

export interface Ability {
  ability: Ability2;
  is_hidden: boolean;
  slot: number;
}

export interface Form {
  name: string;
  url: string;
}

export interface Version {
  name: string;
  url: string;
}

export interface GameIndice {
  game_index: number;
  version: Version;
}

export interface Item {
  name: string;
  url: string;
}

export interface Version2 {
  name: string;
  url: string;
}

export interface VersionDetail {
  rarity: number;
  version: Version2;
}

export interface HeldItem {
  item: Item;
  version_details: VersionDetail[];
}

export interface Move2 {
  name: string;
  url: string;
}

export interface MoveLearnMethod {
  name: string;
  url: string;
}

export interface VersionGroup {
  name: string;
  url: string;
}

export interface VersionGroupDetail {
  level_learned_at: number;
  move_learn_method: MoveLearnMethod;
  version_group: VersionGroup;
}

export interface Move {
  move: Move2;
  version_group_details: VersionGroupDetail[];
}

export interface Species {
  name: string;
  url: string;
}

export interface DreamWorld {
  front_default: string;
  front_female?: never;
}

export interface Home {
  front_default: string;
  front_female: string;
  front_shiny: string;
  front_shiny_female: string;
}

export interface OfficialArtwork {
  front_default: string;
  front_shiny: string;
}

export interface Other {
  dream_world: DreamWorld;
  home: Home;
  "official-artwork": OfficialArtwork;
}
// ...
export interface Versions {
  "generation-i": GenerationI;
  "generation-ii": GenerationIi;
  "generation-iii": GenerationIii;
  "generation-iv": GenerationIv;
  "generation-v": GenerationV;
  "generation-vi": GenerationVi;
  "generation-vii": GenerationVii;
  "generation-viii": GenerationViii;
}

export interface Sprites {
  back_default: string;
  back_female: string;
  back_shiny: string;
  back_shiny_female: string;
  front_default: string;
  front_female: string;
  front_shiny: string;
  front_shiny_female: string;
  other: Other;
  versions: Versions;
}

export interface Stat2 {
  name: string;
  url: string;
}

export interface Stat {
  base_stat: number;
  effort: number;
  stat: Stat2;
}

export interface Type2 {
  name: string;
  url: string;
}

export interface Type {
  slot: number;
  type: Type2;
}

export interface Pokemon {
  slot: number;
  pokemon: Type2;
}

export interface PokemonData {
  abilities: Ability[];
  base_experience?: number;
  forms?: Form[];
  game_indices?: GameIndice[];
  height?: number;
  held_items?: HeldItem[];
  id: number;
  is_default?: boolean;
  location_area_encounters?: string;
  moves: Move[];
  name: string;
  order?: number;
  past_types?: never[];
  species?: Species;
  sprites?: Sprites;
  stats?: Stat[];
  types?: Type[];
  weight?: number;
  pokemon?: Pokemon[];
}

export type PokemonDataApiRes = PokemonData & ErrorHandling;